fixed dbconnect, fixed getserverside props requests, added check if admin do post/delete/put requests

+++ pages/api/products/[id].js
@@ -5,9 +5,12 @@ export default async function handler(req, res) {
 	const {
 		method,
 		query: { id },
+		cookies,
 	} = req
 
-	dbConnect()
+	const token = cookies.token
+
+	await dbConnect()
 
 	if (method === 'GET') {
 		try {
@@ -18,16 +21,10 @@ export default async function handler(req, res) {
 		}
 	}
 
-	// if (method === 'PUT') {
-	// 	try {
-	// 		const product = await Product.create(req.body)
-	// 		res.status(201).json(product)
-	// 	} catch (error) {
-	// 		res.status(500).json(error)
-	// 	}
-	// }
-
 	if (method === 'DELETE') {
+		if (!token || token !== process.env.TOKEN) {
+			return res.status(401).json('Not authenticated!')
+		}
 		try {
 			await Product.findByIdAndDelete(id)
 			res.status(200).json('The product has been deleted')
